Implementing gene/exon groups, SpliceOver calculations
Collates data and implements "SpliceOver" concept
SpliceOver summates all junctions that compete within same "exon island" with the junction being assessed (upstream or downstream in case of junctions, left or right in case of introns)
In unstranded IR, IR and junction groups are different.
In IR, every IR event with same IRgroup combination have been assessed for the same intronic regions, and this in essense is the same IR event (but different alternative junction events).
May be used for downstream filtering purposes