Improve NMD detection efficiency; AHCache d/l from FTP
NMD: first 1000 bases of intron are retrieved; if there is an in-frame stop codon already then the whole intronic sequence is not retrieved. Full-sequence is retrieved from NMD=FALSE introns.

GTF is in gzip compression on Ensembl which is much smaller than that in AnnotationHub. Prefer to get resource from gzipped file